(index, refac) Move domain ranking to a better named package
