+++ code/index/java/nu/marginalia/domainranking/accumulator/RankingResultAccumulator.java
@@ -1,4 +1,4 @@
-package nu.marginalia.ranking.domains.accumulator;
+package nu.marginalia.domainranking.accumulator;
 
 public interface RankingResultAccumulator<T> {
     void add(int domainId, int rank);
