+++ code/index/java/nu/marginalia/domainranking/accumulator/RankingResultListAccumulator.java
@@ -1,4 +1,4 @@
-package nu.marginalia.ranking.domains.accumulator;
+package nu.marginalia.domainranking.accumulator;
 
 import gnu.trove.list.array.TIntArrayList;
