+++ code/index/java/nu/marginalia/domainranking/data/AbstractGraphSource.java
@@ -1,4 +1,4 @@
-package nu.marginalia.ranking.domains.data;
+package nu.marginalia.domainranking.data;
 
 import com.zaxxer.hikari.HikariDataSource;
 import org.jgrapht.Graph;
