+++ code/index/java/nu/marginalia/domainranking/data/GraphSource.java
@@ -1,4 +1,4 @@
-package nu.marginalia.ranking.domains.data;
+package nu.marginalia.domainranking.data;
 
 import org.jgrapht.Graph;
