+++ code/index/java/nu/marginalia/domainranking/data/InvertedLinkGraphSource.java
@@ -1,4 +1,4 @@
-package nu.marginalia.ranking.domains.data;
+package nu.marginalia.domainranking.data;
 
 import com.google.inject.Inject;
 import com.zaxxer.hikari.HikariDataSource;
@@ -7,12 +7,15 @@ import org.jgrapht.Graph;
 import org.jgrapht.graph.DefaultDirectedGraph;
 import org.jgrapht.graph.DefaultEdge;
 
-/** A source for the regular link graph. */
-public class LinkGraphSource extends AbstractGraphSource {
+
+/** A source for the inverted link graph,
+ * which is the same as the regular graph except
+ * the direction of the links have been inverted */
+public class InvertedLinkGraphSource extends AbstractGraphSource {
     private final AggregateLinkGraphClient graphClient;
 
     @Inject
-    public LinkGraphSource(HikariDataSource dataSource, AggregateLinkGraphClient graphClient) {
+    public InvertedLinkGraphSource(HikariDataSource dataSource, AggregateLinkGraphClient graphClient) {
         super(dataSource);
         this.graphClient = graphClient;
     }
@@ -34,11 +37,11 @@ public class LinkGraphSource extends AbstractGraphSource {
                     continue;
                 }
 
-                graph.addEdge(iter.source(), iter.dest());
+                // Invert the edge
+                graph.addEdge(iter.dest(), iter.source());
             }
 
             return graph;
-
         }
         catch (Exception ex) {
             throw new RuntimeException(ex);
