+++ code/index/java/nu/marginalia/index/searchset/SearchSetsService.java
@@ -6,14 +6,14 @@ import gnu.trove.list.TIntList;
 import it.unimi.dsi.fastutil.ints.IntOpenHashSet;
 import nu.marginalia.db.DomainRankingSetsService;
 import nu.marginalia.db.DomainTypes;
+import nu.marginalia.domainranking.PageRankDomainRanker;
+import nu.marginalia.domainranking.accumulator.RankingResultHashMapAccumulator;
+import nu.marginalia.domainranking.accumulator.RankingResultHashSetAccumulator;
+import nu.marginalia.domainranking.data.GraphSource;
+import nu.marginalia.domainranking.data.LinkGraphSource;
+import nu.marginalia.domainranking.data.SimilarityGraphSource;
 import nu.marginalia.index.IndexFactory;
 import nu.marginalia.index.domainrankings.DomainRankings;
-import nu.marginalia.ranking.domains.PageRankDomainRanker;
-import nu.marginalia.ranking.domains.accumulator.RankingResultHashMapAccumulator;
-import nu.marginalia.ranking.domains.accumulator.RankingResultHashSetAccumulator;
-import nu.marginalia.ranking.domains.data.GraphSource;
-import nu.marginalia.ranking.domains.data.LinkGraphSource;
-import nu.marginalia.ranking.domains.data.SimilarityGraphSource;
 import nu.marginalia.service.control.ServiceEventLog;
 import nu.marginalia.service.module.ServiceConfiguration;
 import org.slf4j.Logger;
